# web/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Books API",
        "description": "Пример простого REST API",
        "contact": {
            "email": "iagmail@mail.ru"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Swagger Yii2 API Example (built in) - пример простого API"
        }
    ],
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Вход пользователя",
                "description": "Вход пользователя",
                "operationId": "login",
                "requestBody": {
                    "description": "Вход пользователя.<br/>\n     *                      Для входа использовать: admin/admin<br/>\n     *                      см. также Insomnia, тот же метод (после импорта данных)<br/>\n     *                      Изменить время жизни access токена см. config/params.php, параметр accessTokenTime,<br/>\n     *                      (также TTL refresh токена - параметр refreshTokenTime, там же)",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "Имя пользователя"
                                    },
                                    "password": {
                                        "description": "Пароль"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Вход - успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Токены",
                                    "properties": {
                                        "accessToken": {
                                            "description": "Access токен",
                                            "type": "string"
                                        },
                                        "refreshToken": {
                                            "description": "Refresh токен",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не прошел аутентификацию",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибки валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTAuthentification": []
                    }
                ]
            }
        },
        "/api/v1/auth/refresh-tokens": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Обновление токенов",
                "description": "Обновление токенов",
                "operationId": "refresh",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refreshToken"
                                ],
                                "properties": {
                                    "refreshToken": {
                                        "description": "Refresh токен",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Токены обновлены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "accessToken": {
                                            "type": "string"
                                        },
                                        "refreshToken": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Токен отсутствует или не найден на сервере",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Время жизни токена истекло",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь (или токен) не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTAuthentification": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "patch": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выход пользователя",
                "description": "Выход пользователя",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Выход - успешно"
                    },
                    "400": {
                        "description": "Токен отсутствует или структура токена некорректна",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Время жизни токена истекло",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTAuthentification": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Exception": {
                "title": "Exception",
                "description": "Данные при возникновении исключения",
                "properties": {
                    "name": {
                        "description": "Наименование исключения",
                        "type": "string"
                    },
                    "message": {
                        "description": "Текст ошибки",
                        "type": "string"
                    },
                    "code": {
                        "description": "Код",
                        "type": "integer",
                        "enum": [
                            10,
                            20,
                            30,
                            40,
                            50,
                            60,
                            70
                        ]
                    },
                    "status": {
                        "description": "HTTP статус",
                        "type": "integer"
                    },
                    "type": {
                        "description": "Тип исключения",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Error": {
                "properties": {
                    "field": {
                        "description": "Поле",
                        "type": "string"
                    },
                    "message": {
                        "description": "Текст ошибки",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "JWTAuthentification": {
                "type": "http",
                "description": "Authentification by JWT token",
                "name": "JWT",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}